moved animated folder to components folder
