+++ frontend/app/src/sphinxUI/modal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import FadeLeft from '../animated/fadeLeft';
+import FadeLeft from '../components/animated/fadeLeft';
 import { IconButton } from '.';
 import { colors } from '../colors';
 import { Portal } from './portal';
